*Refactored*Added component modules for future build*introduced first route
