--- src/App.css
@@ -1,51 +1,0 @@
-.App {
-  background-color: #282c34;
-}
-
-.App-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: white;
-}
-
-.card-parent-div {
-  display: flex;
-  justify-content: space-evenly;
-}
-.card {
-  text-align: center;
-  border: 4px solid orange;
-  padding: 16px 8px 8px 8px;
-  border-radius: 12px;
-  color: white;
-}
-
-@keyframes up-bump {
-  0% {
-    padding-top: 2em;
-  }
-  50% {
-    padding-top: 1.5em;
-  }
-  100% {
-    padding-top: 2em;
-  }
-}
-
-.card:hover {
-  animation: up-bump 0.4s ease;
-}
-
-.link-theme {
-  color: white;
-}
-.link:visited {
-  color: none;
-}
