+++ src/components/homepage/HomePage.css
@@ -1,0 +1,22 @@
+.card-parent-div {
+  display: flex;
+  justify-content: space-evenly;
+}
+.card {
+  text-align: center;
+  border: 4px solid orange;
+  padding: 16px 8px 8px 8px;
+  border-radius: 12px;
+  color: white;
+}
+
+.card:hover {
+  animation: up-bump 0.4s ease;
+}
+
+.link-theme {
+  color: white;
+}
+.link:visited {
+  color: none;
+}
